Moved class name calculations into a util function
